requirements.txt
azure-core==1.20.1
azure-cosmos==4.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
